+++ src/main/groovy/io/seqera/wave/storage/ManifestCacheStore.groovy
@@ -25,8 +25,8 @@ import groovy.util.logging.Slf4j
 import io.micronaut.context.annotation.Value
 import io.seqera.wave.api.ContainerLayer
 import io.seqera.wave.encoder.MoshiEncodeStrategy
-import io.seqera.wave.store.state.AbstractCacheStore
-import io.seqera.wave.store.state.impl.CacheProvider
+import io.seqera.wave.store.state.AbstractStateStore
+import io.seqera.wave.store.state.impl.StateProvider
 import jakarta.inject.Singleton
 /**
  * Implements manifest cache for {@link DigestStore}
@@ -36,12 +36,12 @@ import jakarta.inject.Singleton
 @Slf4j
 @Singleton
 @CompileStatic
-class ManifestCacheStore extends AbstractCacheStore<DigestStore> implements Storage {
+class ManifestCacheStore extends AbstractStateStore<DigestStore> implements Storage {
 
     final private Duration duration
 
     ManifestCacheStore(
-            CacheProvider<String, String> provider,
+            StateProvider<String, String> provider,
             @Value('${wave.storage.cache.duration:`1h`}') Duration duration)
     {
         super(provider, new MoshiEncodeStrategy<DigestStore>() {})
