+++ src/test/groovy/io/seqera/wave/service/builder/ContainerBuildServiceTest.groovy
@@ -95,7 +95,7 @@ class ContainerBuildServiceTest extends Specification {
     @Inject ContainerInspectServiceImpl dockerAuthService
     @Inject HttpClientConfig httpClientConfig
     @Inject BuildConfig buildConfig
-    @Inject BuildCacheStore buildCacheStore
+    @Inject BuildStoreImpl buildCacheStore
     @Inject PersistenceService persistenceService
     @Inject JobService jobService
